# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CommerceVision Control API",
    "summary": "Control plane for durable ecommerce visual workflows",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/workflows": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Create Workflow",
        "operationId": "create_workflow_api_v1_workflows_post",
        "parameters": [
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "X-Workspace-Id"
            }
          },
          {
            "name": "X-Actor-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "X-Actor-Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 256,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "List Workflows",
        "operationId": "list_workflows_api_v1_workflows_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Get Workflow",
        "operationId": "get_workflow_api_v1_workflows__workflow_id__get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}:cancel": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Cancel Workflow",
        "operationId": "cancel_workflow_api_v1_workflows__workflow_id__cancel_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "X-Workspace-Id"
            }
          },
          {
            "name": "X-Actor-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "X-Actor-Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 256,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCancelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}/events": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Workflow Events",
        "operationId": "workflow_events_api_v1_workflows__workflow_id__events_get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventResponse"
                  },
                  "title": "Response Workflow Events Api V1 Workflows  Workflow Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}/creative-plan:approve": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Approve Plan",
        "operationId": "approve_plan_api_v1_workflows__workflow_id__creative_plan_approve_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "X-Workspace-Id"
            }
          },
          {
            "name": "X-Actor-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "X-Actor-Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 256,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}/creative-plan:reject": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Reject Plan",
        "operationId": "reject_plan_api_v1_workflows__workflow_id__creative_plan_reject_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "X-Workspace-Id"
            }
          },
          {
            "name": "X-Actor-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "X-Actor-Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 256,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}/results:approve": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Approve Results",
        "operationId": "approve_results_api_v1_workflows__workflow_id__results_approve_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "X-Workspace-Id"
            }
          },
          {
            "name": "X-Actor-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "X-Actor-Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 256,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}/results:regenerate": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Regenerate Results",
        "operationId": "regenerate_results_api_v1_workflows__workflow_id__results_regenerate_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "X-Workspace-Id"
            }
          },
          {
            "name": "X-Actor-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "X-Actor-Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 256,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApprovalRequest": {
        "properties": {
          "expected_workflow_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Workflow Version"
          },
          "subject_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Subject Id"
          },
          "subject_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Subject Version"
          },
          "decision": {
            "$ref": "#/components/schemas/ApprovalDecision"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "comment_ref": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment Ref"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_workflow_version",
          "subject_id",
          "subject_version",
          "decision"
        ],
        "title": "ApprovalRequest"
      },
...
      "EventResponse": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "schema_version": {
            "type": "integer",
            "title": "Schema Version"
          },
          "aggregate_type": {
            "type": "string",
            "title": "Aggregate Type"
          },
          "aggregate_id": {
            "type": "string",
            "title": "Aggregate Id"
          },
          "aggregate_version": {
            "type": "integer",
            "title": "Aggregate Version"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "event_type",
          "schema_version",
          "aggregate_type",
          "aggregate_id",
          "aggregate_version",
          "occurred_at",
          "trace_id",
          "payload"
        ],
        "title": "EventResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RetentionStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "EXPIRING",
          "DELETING",
          "EXPIRED"
        ],
        "title": "RetentionStatus"
      },
...
      "WorkflowAttemptResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "step_id": {
            "type": "string",
            "title": "Step Id"
          },
          "attempt_number": {
            "type": "integer",
            "title": "Attempt Number"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "status": {
            "$ref": "#/components/schemas/AttemptStatus"
          },
          "provider_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Request Id"
          },
          "result_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Ref"
          },
          "result_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Data"
          },
          "error_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Class"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "step_id",
          "attempt_number",
          "idempotency_key",
          "status",
          "provider_request_id",
          "result_ref",
          "result_data",
          "error_class",
          "error_message",
          "started_at",
          "completed_at"
        ],
        "title": "WorkflowAttemptResponse"
      },
      "WorkflowCancelRequest": {
        "properties": {
          "expected_workflow_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Workflow Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_workflow_version"
        ],
        "title": "WorkflowCancelRequest"
      },
      "WorkflowCreateRequest": {
        "properties": {
          "workflow_type": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Workflow Type",
            "default": "FIXTURE_IMAGE_GENERATION"
          },
          "input_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Data"
          },
          "retention_hours": {
            "type": "integer",
            "maximum": 168.0,
            "minimum": 1.0,
            "title": "Retention Hours",
            "default": 72
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "WorkflowCreateRequest"
      },
      "WorkflowListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkflowResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "WorkflowListResponse"
      },
      "WorkflowResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "workflow_type": {
            "type": "string",
            "title": "Workflow Type"
          },
          "status": {
            "$ref": "#/components/schemas/WorkflowStatus"
          },
          "retention_status": {
            "$ref": "#/components/schemas/RetentionStatus"
          },
          "current_node": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Node"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "input_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Data"
          },
          "result_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Data"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "cancellation_requested_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancellation Requested At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/WorkflowStepResponse"
            },
            "type": "array",
            "title": "Steps"
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/WorkflowAttemptResponse"
            },
            "type": "array",
            "title": "Attempts"
          },
          "approvals": {
            "items": {
              "$ref": "#/components/schemas/ApprovalResponse"
            },
            "type": "array",
            "title": "Approvals"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "created_by",
          "workflow_type",
          "status",
          "retention_status",
          "current_node",
          "version",
          "input_data",
          "result_data",
          "expires_at",
          "cancellation_requested_at",
          "created_at",
          "updated_at"
        ],
        "title": "WorkflowResponse"
      },
      "WorkflowStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "INGESTING",
          "UNDERSTANDING",
          "AWAITING_PRODUCT_CONFIRMATION",
          "RETRIEVING",
          "PLANNING",
          "AWAITING_PLAN_APPROVAL",
          "GENERATING",
          "EVALUATING",
          "REPAIRING",
          "AWAITING_RESULT_APPROVAL",
          "EXPORTING",
          "COMPLETED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "WorkflowStatus"
      },
...
    }
  }
}